feat: payment with stripe webhook

+++ src/app/modules/docotorSchedule/docotorSchedule.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import status from "http-status";
+import AppError from "../../errorHelper/AppError";
 import { IQueryParams } from "../../interfaces/query.interface";
 import { catchAsync } from "../../shared/catchAsync";
 import { sendResponse } from "../../shared/sendResponse";
@@ -10,6 +11,11 @@ export const DoctorScheduleController = {
   createMyDoctorSchedule: catchAsync(async (req: Request, res: Response) => {
      const payload = req.body;
     const user = req.user;
+
+    if (!user) {
+      throw new AppError(status.UNAUTHORIZED, "Unauthorized access");
+    }
+
     const doctorSchedule = await doctorScheduleService.createMyDoctorSchedule(user, payload);
 
     sendResponse({
@@ -23,7 +29,11 @@ export const DoctorScheduleController = {
 
   getMyDoctorSchedules: catchAsync(async (req: Request, res: Response) => {
     const user = req.user;
-    const query = req.query ;
+    const query = req.query as IQueryParams;
+
+    if (!user) {
+      throw new AppError(status.UNAUTHORIZED, "Unauthorized access");
+    }
 
     const result = await doctorScheduleService.getMyDoctorSchedules(user, query);
 
@@ -67,6 +77,10 @@ export const DoctorScheduleController = {
   updateMyDoctorSchedule: catchAsync(async (req: Request, res: Response) => {
     const user = req.user;
 
+    if (!user) {
+      throw new AppError(status.UNAUTHORIZED, "Unauthorized access");
+    }
+
     const result = await doctorScheduleService.updateMyDoctorSchedule(user, req.body);
 
     sendResponse({
@@ -81,6 +95,10 @@ export const DoctorScheduleController = {
   deleteMyDoctorSchedule: catchAsync(async (req: Request, res: Response) => {
     const { id } = req.params;
     const user = req.user;
+
+    if (!user) {
+      throw new AppError(status.UNAUTHORIZED, "Unauthorized access");
+    }
 
     await doctorScheduleService.deleteMyDoctorSchedule(id as string, user);
